Move writeable_site_required to common.views
