+++ pouet/views.py
@@ -1,8 +1,8 @@
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, render
-from read_only_mode import writeable_site_required
 
+from common.views import writeable_site_required
 from demoscene.models import Edit, Releaser
 from pouet.matching import get_match_data
 from pouet.models import GroupMatchInfo
